+++ builds/4chan-X.js
@@ -1,6 +1,6 @@
 // ==UserScript==
 // @name         4chan X
-// @version      1.1.6
+// @version      1.1.7
 // @namespace    4chan-X
 // @description  Cross-browser userscript for maximum lurking on 4chan.
 // @license      MIT; https://github.com/seaweedchan/4chan-x/blob/master/LICENSE 
@@ -18,7 +18,7 @@
 // @icon         data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADAAAAAwAgMAAAAqbBEUAAAACVBMVEUAAGcAAABmzDNZt9VtAAAAAXRSTlMAQObYZgAAAHFJREFUKFOt0LENACEIBdBv4Qju4wgWanEj3D6OcIVMKaitYHEU/jwTCQj8W75kiVCSBvdQ5/AvfVHBin11BgdRq3ysBgfwBDRrj3MCIA+oAQaku/Q1cNctrAmyDl577tOThYt/Y1RBM4DgOHzM0HFTAyLukH/cmRnqAAAAAElFTkSuQmCC
 // ==/UserScript==
 /*
-* 4chan X - Version 1.1.6 - 2013-05-01
+* 4chan X - Version 1.1.7 - 2013-05-01
 *
 * Licensed under the MIT license.
 * https://github.com/seaweedchan/4chan-x/blob/master/LICENSE
@@ -308,7 +308,7 @@
   doc = d.documentElement;
 
   g = {
-    VERSION: '1.1.6',
+    VERSION: '1.1.7',
     NAMESPACE: '4chan X.',
     boards: {},
     threads: {},