+++ CHANGELOG.md
@@ -2,6 +2,7 @@ seaweedchan:
 - External image embedding
 - Account for time options in youtube links for embedding
 - Once again remove /v/ and /vg/ archiving... ;_;
+- Add paste.installgentoo.net embedding
 
 ### 1.1.6 - 2013-05-01
 seaweedchan: