+++ CHANGELOG.md
@@ -4,6 +4,7 @@ seaweedchan:
 - Once again remove /v/ and /vg/ archiving... ;_;
 - Add paste.installgentoo.com embedding
 - Added `Posting Success Notifications` option to make "Post Successful!" and "_____ uploaded" notifications optional
+- Added `Allow False Positives` option under Linkification, giving the user more control over what's linkified.
 
 ### 1.1.6 - 2013-05-01
 seaweedchan: