+++ CHANGELOG.md
@@ -5,6 +5,7 @@ seaweedchan:
 - Add paste.installgentoo.com embedding
 - Added `Posting Success Notifications` option to make "Post Successful!" and "_____ uploaded" notifications optional
 - Added `Allow False Positives` option under Linkification, giving the user more control over what's linkified.
+- Fix URL for update checking
 
 ### 1.1.6 - 2013-05-01
 seaweedchan: